[refactor] LottoTicket package 변경
