+++ src/main/java/lotto/service/domain/LottoTicketGenerator.java
@@ -1,7 +1,5 @@
 package lotto.service.domain;
 
-import lotto.service.model.LottoTicket;
-
 public interface LottoTicketGenerator {
     LottoTicket makeLottoTicket();
 }
